Extract source product analytics router
