+++ quality/baseline_report.md
@@ -10,9 +10,9 @@ documentation posture, and validation gates. It is not a completion claim.
 
 ## Code Size Baseline
 
-- Python source files under `src/`: 54
+- Python source files under `src/`: 55
 - Python test files under `tests/`: 77
-- API entry point route/middleware/handler decorators in `src/app/main.py`: 8
+- API entry point route/middleware/handler decorators in `src/app/main.py`: 5
 
 ### Largest Python Files
 
@@ -81,10 +81,11 @@ documentation posture, and validation gates. It is not a completion claim.
    middleware registration, and risk analytics route orchestration. Standard OpenAPI error metadata
    and exception-handler registration now live in `src/app/api_errors.py`; health, readiness,
    metrics, operational diagnostics, trust telemetry, and capability publication now live in
-   `src/app/routers/operational.py`.
-2. Operational and capability routes are split into a router module. Risk analytics routes are not
-   yet split into route modules; their OpenAPI metadata and dependency resolution remain coupled to
-   the FastAPI app instance.
+   `src/app/routers/operational.py`; stateless source-product endpoints now live in
+   `src/app/routers/source_products.py`.
+2. Operational, capability, and stateless source-product routes are split into router modules.
+   Core calculation routes are not yet split into route modules; their OpenAPI metadata and
+   dependency resolution remain coupled to the FastAPI app instance.
 3. Business calculations already live mostly under `src/app/services`, which gives the next slices
    a workable extraction boundary.
 4. Infrastructure clients already sit under `src/app/integrations`, but route handlers still
@@ -144,7 +145,7 @@ tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matri
 tests/unit/test_upstream_errors.py::test_invalid_payload_and_missing_data_carry_structured_categories
 tests/unit/test_upstream_errors.py::test_extract_upstream_error_detail_variants
 
-369 tests collected in 1.50s
+369 tests collected in 2.47s
 ```
 - Import-linter report-only: reported exit 127
