+++ scripts/generate_quality_baseline.py
@@ -152,10 +152,11 @@ documentation posture, and validation gates. It is not a completion claim.
    middleware registration, and risk analytics route orchestration. Standard OpenAPI error metadata
    and exception-handler registration now live in `src/app/api_errors.py`; health, readiness,
    metrics, operational diagnostics, trust telemetry, and capability publication now live in
-   `src/app/routers/operational.py`.
-2. Operational and capability routes are split into a router module. Risk analytics routes are not
-   yet split into route modules; their OpenAPI metadata and dependency resolution remain coupled to
-   the FastAPI app instance.
+   `src/app/routers/operational.py`; stateless source-product endpoints now live in
+   `src/app/routers/source_products.py`.
+2. Operational, capability, and stateless source-product routes are split into router modules.
+   Core calculation routes are not yet split into route modules; their OpenAPI metadata and
+   dependency resolution remain coupled to the FastAPI app instance.
 3. Business calculations already live mostly under `src/app/services`, which gives the next slices
    a workable extraction boundary.
 4. Infrastructure clients already sit under `src/app/integrations`, but route handlers still
